Fixup Windows platform for UTF8. * Added Widen and Narrow functions. UTF8 <-> UTF16 conversion. C++11. * Fix OpenGL samples.

+++ gwen/include/Gwen/Events.h
@@ -41,7 +41,7 @@ namespace Gwen
             Gwen::String String;
             int Integer;
             float Float;
-            uint64_t UnsignedLongLong;
+            unsigned long long UnsignedLongLong;
         };
